Add files via upload
This game was created by: Mohamed, Jeferson, Andrea, MD Shamiul. 

Fun Game is a text-based (also known as interactive fiction) adventure game, where all the interaction takes place through on-screen words. The player enters a maze, which is described in the narrative and the player submits simple commands to interact with the game world. The purpose of the game is that the player should survive and find their way through the maze to safety.
The game world has several different areas, such as treasure room, graveyard, crossroad, etc. The player can move through these environments in a certain order.
There are various types of items in the game that the player can collect through their journey. Having these items in the player’s inventory effects the gameplay.
There are two outcomes: the player can die if they make a bad decision, but with the correct choices they can reach the exit of the maze.